+++ faker_biology/taxonomy/taxonomy_data.py
@@ -1,3 +1,6 @@
+from collections import namedtuple
+
+## these 2 lists should be equal length
 MODEL_ORGANISMS_ENGLISH = [
     "Fly",
     "Mouse",
@@ -20,3 +23,13 @@ MODEL_ORGANISMS_LATIN = [
     "Saccharomyces cervisiae",
     "Danio Rerio",
 ]
+
+MODEL_ORGANISM = namedtuple("model_organism", ["english", "latin"])
+
+zipped = zip(MODEL_ORGANISMS_ENGLISH, MODEL_ORGANISMS_LATIN)
+list_zipped = list(zipped)
+MODEL_ORGANISMS = [MODEL_ORGANISM(*mo) for mo in list_zipped]
+
+
+
+